% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{game}
\alias{game}
\title{Arranca el silly game}
\usage{
game()
}
\description{
El silly game es una aventura conversacional. Tú eres un ser anónimo cuyo
objetivo es escapar de una celda. Tu principal misión es abrir la puerta.
}
\details{
Para jugar, debes escribir órdenes para tu personaje. Algunos ejemplos de
órdenes son:
\preformatted{
>> explorar
>> mirar puerta
>> abrir puerta
>> salir
}
}
\section{Salir}{

Si quieres salir del juego antes de que este termine, simplemente escribe:
\preformatted{
>> q
}
}

